Silence/fix double to float conversion warnings

+++ pcbnew/pcb_parser.cpp
@@ -365,7 +365,7 @@ void PCB_PARSER::parseEDA_TEXT( EDA_TEXT* aText )
     // 60mils is 1.524mm
     if( !foundTextSize )
     {
-    	const float defaultTextSize = 1.524f * IU_PER_MM;
+    	const double defaultTextSize = 1.524 * IU_PER_MM;
 
     	aText->SetTextSize( wxSize( defaultTextSize, defaultTextSize ) );
     }
